requirements.txt
beautifulsoup4
SQLAlchemy
python-telegram-bot
